refactor seperate functions into new files and add basic layout styling
