+++ app/javascript/components/EditForm.js
@@ -4,16 +4,11 @@ import { map } from "lodash"
 import { Form, Field } from "react-final-form";
 import 'isomorphic-fetch';
 
+import { incomeRows, customerDetailFields } from '../helpers/data'
 import { getCSRF } from '../helpers/getCSRF';
-const databaseURL = process.env.APP_URL  
+import { SingleInput, RadioInput, TableInput } from './inputs'
 
-const Error = ({ name }) => (
-  <Field name={name} subscription={{ error: true, touched: true }}>
-    {({ meta: { error, touched } }) =>
-      error && touched ? <span>{error}</span> : null
-    }
-  </Field>
-);
+const databaseURL = process.env.APP_URL  
 
 const Condition = ({ when, is, children }) => (
   <Field name={when} subscription={{ value: true }}>
@@ -44,6 +39,7 @@ class EditRebateForm extends React.Component {
     const { fields } = this.props
     return (
       <Form
+        className="rebate-wrapper"
         onSubmit={this.onSubmit.bind(this)}
         initialValues={this.props.fields}
         validate={values => {
@@ -63,29 +59,31 @@ class EditRebateForm extends React.Component {
       >
         {({ handleSubmit, reset, submitting, pristine, values }) => (
           <form onSubmit={handleSubmit}>
-            {map(customerDetailFields, (field) => {
-              return field.type == 'radio'
-              ? RadioInput(field)
-              : SingleInput(field)
-            })}
+            <div className="flex-row">
+              {map(customerDetailFields, (field) => {
+                return field.type == 'radio'
+                ? RadioInput(field)
+                : SingleInput(field)
+              })}
+            </div>
             <div>
               <div className="flex-column">
                 <h2 className="flex-item">Total Income</h2>
                 <div className="flex-row">
-                  <h2 className="flex-item" >Income Type</h2>
-                  <h2 className="flex-item" >Applicant</h2>
-                  <h2 className="flex-item" >Partner</h2>
+                  <h2 className="flex-item one-third" >Income Type</h2>
+                  <h2 className="flex-item one-third" >Applicant</h2>
+                  <h2 className="flex-item one-third" >Partner</h2>
                 </div>
                 {map(incomeRows, (field) => {
                 return (
-                  <div className="flex-row">
-                    <div className="flex-item"> 
+                  <div key={field.id} className="flex-row">
+                    <div className="flex-item one-third"> 
                       {field.label}
                     </div>
-                    <div className="flex-item">
+                    <div className="flex-item one-third">
                       {TableInput({...field, id: `${field.id}-Applicant`})}
                     </div>
-                    <div className="flex-item">
+                    <div className="flex-item one-third">
                       {TableInput({...field, id: `${field.id}-Partner`})}
                     </div>
                   </div>
@@ -109,134 +107,4 @@ class EditRebateForm extends React.Component {
   }
 }
 
-EditRebateForm.propTypes = {};
-
-const incomeRows = [
-  {
-    id: "other_superannuation",
-    label: "Other superannuation"
-  },
-  {
-    id: "interest_dividends",
-    label: "Interest / Dividends"
-  },
-  {
-    id: "wages_salery",
-    label: "Wages or Salary"
-  }
-
-]   
-const customerDetailFields = [
-  {
-    id:"full_name",
-    label: "Name",
-  },
-  {
-    id:"customer_id",
-    label: "Customer ID (optional)",
-    placeholder: "0000000",
-  },
-  {
-    id:"address",
-    label: "Address",
-    fullWidth: true
-  },
-  {
-    id:"total_rates",
-    label: "Total Rates",
-  },
-  {
-    id:"valuation_id",
-    label: "Valuation ID (optional)",
-    placeholder: "00000 000 00",
-  },
-  {
-    id: 'moved_within_rating_year',
-    label: "Moved within rating year?",
-    type:"radio"
-  },
-  {
-    id: 'lived_in_property_1_July',
-    label: "Lived in property 1 July 2018",
-    type:"radio"
-  },
-  {
-    id:"email",
-    label: "Email",
-    type:"email",
-  },
-  {
-    id:"phone_number",
-    label: "Phone",
-    type:"tel",
-  },
-  {
-    id: 'spouse_or_partner',
-    label: "Spouse or Partner",
-    type:"radio"
-  },
-  {
-    id:"dependents",
-    label: "Dependents",
-    type:"number",
-    placeholder: "0",
-  },
-  {
-    id:"50%_claimed_expenses",
-    label: "50% claimed as expenses?",
-    type:"radio"
-  },
-  {
-    id:"occupation",
-    label: "Occupation",
-  },
-]
-
-const SingleInput = ({ id, label, placeholder, type = "text" }) => {
-  return (
-  <div key={id} className=''>
-    <label>{label}</label>
-    <Field
-      name={id}
-      label= {label}
-      placeholder= {placeholder}
-      component='input'
-      type={type}
-    />
-    <Error name={id} />
-  </div>
-)}
-const TableInput = ({ id, label, placeholder, type = "text" }) => {
-  return (
-  <div key={id} className=''>
-    <Field
-      name={id}
-      label= {label}
-      placeholder= {placeholder}
-      component='input'
-      type={type}
-    />
-    <Error name={id} />
-  </div>
-)}
-
-const RadioInput = ({ id, label, type }) => (
-  <div key={id} >
-    <label>{label}</label>
-    <div>
-      {map(["yes", "no"], value =>
-      <label key={`${id}-${value}`} >
-        <Field
-          name={id}
-          component="input"
-          type={type}
-          value={value}
-        />{" "}
-        {value}
-      </label>
-      )}
-      <Error name={id} />
-    </div>
-  </div>
-)
 export default EditRebateForm
