+++ app/views/admin/rebate_forms/edit.html.haml
@@ -1,3 +1,4 @@
-%h1 Edit Application
+%h1 Customer Details
+%h2 Signature required
 
 = react_component("EditForm", { fields: @rebate_form.fields, valuation: @rebate_form.valuation_id })
